* C module: C.LinkPrebuiltLibraries updates with support for relative paths, libname.* linking on GCC based platforms, and proper dependency checking on Windows.
Ben Hymers reported (http://jamplus.org/boards/1/topics/183) that C.LinkPrebuiltLibraries was not working smoothly on Linux.  The only way to link .dylib/.so files was to use C.LinkFlags with the -l option.  He suggested some approaches to fixing this and provided implementation details.

For GCC-based platforms, this changelist uses Ben's advice and builds on it further.  Libraries passed to C.LinkPrebuiltLibraries in the form of 'name' (without the quotes, not prefaced with 'lib', and not containing an extension or directory path) are passed to the linker with the -l option.  Proper dependency checking is not performed on these style of libraries at this time.

Libraries passed to C.LinkPrebuiltLibraries with a directory (relative or absolute) or extension are resolved to their proper location on disk and passed to the linker directory.