C module: Add support for a TOOLCHAIN command line parameter
If the TOOLCHAIN command line parameter is specified, then that named
toolchain is used for the build.  To use the tcc-win32-debug toolchain,
run Jam like:

    jam TOOLCHAIN=tcc-win32-debug

The default toolchain is dependent on the platform.  On Windows, it
defaults to win32-release with an automatic detection of the appropriate
compiler.  Internally, it is handled as if the command-line had stated:

    jam TOOLCHAIN=win32-release

The PLATFORM and CONFIG command line parameters still work.  If the
command line states:

    jam PLATFORM=win32 CONFIG=release

then it is as if the following TOOLCHAIN were specified:

    jam TOOLCHAIN=win32-release