Ensure missing targets which depend on MightNotUpdate targets are created.
Currently the complicated checks in the case for MightNotUpdate children end up
forcing the parent to a stable fate when the child has not updated, even when
the parent is bound to a missing file.

This work-around avoids the special handling for MightNotUpdate children when
the parent is bound to a missing file.