Allow C.Provision to be called with a filename. The existing behaviour (search for profile based on ID) still functions, but initially it will check if the profile is being passed by filename. We have found this much better, as we now store our profiles in our code repository, and then call C.Provision like:
C.Provision OurGame : "$(Root)/iOS/Provisions/OurGameDev.mobileprovision" ;

This means when the profile needs updating (new device, feature, expired, etc.) we can simply save the new profile into the above location and everything continues working. Previously we had to exact the PROFILE_ID from the profile, and then update the Jamfile.

Also, this fixes an issue where the embedded provision wouldn't update (inside the bundle) when the profile changed (the time stamp check reported the file inside the bundle was newer, the problem is it was newer but the wrong file).