JamToWorkspace: Fix Xcode generation when configurations are not available
When generating a workspace for multiple platforms, some platforms may
have configurations not available on other platforms (i.e. adhoc/
distribution for iOS/OSX aren't available for Tizen). This previously
caused the workspace generation to fail with:

-------------------------------------------------------------------------------
-- ...bin/scripts/ide/xcode.lua:182: attempt to index field '?' (a nil value):
-------------------------------------------------------------------------------
stack traceback:
	...usNextGen/bin/macosx32/../scripts/JamToWorkspace.lua:121: in function <...usNextGen/bin/macosx32/../scripts/JamToWorkspace.lua:97>
	.../bin/scripts/ide/xcode.lua:182: in function 'XcodeHelper_GetProjectExportInfo'
	.../bin/scripts/ide/xcode.lua:582: in function 'Write'
	.../bin/macosx32/../scripts/JamToWorkspace.lua:554: in function 'DumpProject'
	.../bin/macosx32/../scripts/JamToWorkspace.lua:664: in function 'DumpWorkspace'
	.../bin/macosx32/../scripts/JamToWorkspace.lua:988: in function <.../bin/macosx32/../scripts/JamToWorkspace.lua:702>
	[C]: in function 'xpcall'
	.../bin/macosx32/../scripts/JamToWorkspace.lua:1066: in main chunk
	[C]: ?
--------------------------------------------------------------------------------------------------

This commit if checks to avoid accessing nil which resolves the issue (I
appreciate there are some higher level issues about what should happen for
an 'adhoc' 'tizen' build, but regardless this commit hopefully will be seen
as an improvement on the current situation).