Use a global depcache instead of file-level depcaches
It is error prone to assign targets to file-level depcaches and then to
ensure that target always refers to the same depcache and doesn't get
mixed into another one.

This commit still tracks depcaches individually for reads and writes, but
the data storage in memory now allows only one boundname globally instead
of the possibility of the same boundname being referred to in differing
depcache hashes.