Prevent UseCommandLine source targets from being excluded improperly from the parent
This happens when building, say, a precompiled header file that depends on
a .cpp file. A global #define is added. The .cpp file should be built into
the precompiled header file, but instead, the .cpp is removed from the
action's sources due to the 'updated' flag on the action. The new pch file
is not built, and lots of stuff breaks.

Why is this code here? Why was it ever here? It has been too many years,
and I simply do not remember.