Split checksums out of depcache files and into a separate .jamchecksums file
This accounts for uses where there are multiple depcache files that store
information about the same files. Without a shared checksum cache, files
can be rebuilt that do not need to be, but the information to make that
determination is not present in the active depcache file but is, instead,
in another depcache file.