C: Rearrange VC dll dependency chain to account for checksum support
The .lib depends on the .dll for purposes of users of the .dll. However,
the .lib and the .dll are built at the same time, and this makes it
impossible for a checksum to be generated properly. When the .dll hasn't
been built yet, the checksum can't be generated at all. When the .dll has
been built before, the checksum is generated against the existing .dll,
and a new .dll will get created for this run of Jam. When Jam is run again,
the checksum support notes that the .dll was updated and generates a new
checksum.

In order to avoid a never ending build cycle, an 'inner' .lib target is
used to build the shared library, and the public .lib depends on the
'inner' .lib and the .dll. The public .lib target can't create a checksum
until both the 'inner' .lib and the .dll are fully built.