C: Remove dependency of the .obj on the .pch file for VC
Both pch.obj and pch.pch are generated from the same action. As such,
checksums cannot be generated properly for the pch.obj or pch.pch files
until the entire action has finished.

For some reason not apparent to me now, the pch.obj depends on the
pch.pch files. Removing this dependency fixes the checksum issues and
still appears to do the Right Thing for VC++ precompiled headers.