Jambase: Solve a MkDir race condition on Windows resulting randomly in failure
In a multiprocess build, it is possible for the "if not exist DIR" to
be true, but by the time the $(MKDIR) is run, the directory exists from
another path.

This can be reproduced with the following Jamfile.jam and by adding a
"timeout #" to the original MkDir1 before this commit.

Jamfile.jam:

DEPCACHE.standard = .depcache ;

Depends all : b.txt ;
local OUTPUT = "output text b" ;
UseCommandLine b.txt : $(OUTPUT) ;
MakeLocate b.txt : output/a/b ;
CONTENTS on b.txt = $(OUTPUT) ;
Clean clean : b.txt ;
WriteFile b.txt ;

Depends all : a.txt ;
OUTPUT = "output text a" ;
UseCommandLine a.txt : $(OUTPUT) ;
MakeLocate a.txt : output/a ;
CONTENTS on a.txt = $(OUTPUT) ;
Clean clean : a.txt ;
WriteFile a.txt ;